admin: fold synchronizer module into sync command module
