+++ admin/src/lib.rs
@@ -14,7 +14,6 @@ pub mod list_versions;
 pub mod lock;
 pub mod osv_export;
 pub mod prelude;
-pub mod synchronizer;
 pub mod web;
 
 use std::{collections::BTreeMap as Map, error::Error as StdError};
